requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
Flask==2.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
